+++ packages/falcon-ecommerce-uikit/src/SignIn/SignInForm.tsx
@@ -25,7 +25,7 @@ export const SignInForm: React.SFC<SignInFormProps> = ({ onCompleted, id }) => (
         {() => (
           <Form id={id} i18nId="signIn">
             <FormField name="email" type="email" required autoComplete="email" />
-            {/* <FormField
+            <FormField
               name="password"
               type="password"
               // pass empty array, so default password strength validator does not get triggered
@@ -34,15 +34,15 @@ export const SignInForm: React.SFC<SignInFormProps> = ({ onCompleted, id }) => (
               autoComplete="current-password"
             >
               {({ field }) => <PasswordRevealInput {...field} />}
-            </FormField> */}
-            {/* <FlexLayout justifyContent="space-between" alignItems="center" mt="md">
+            </FormField>
+            <FlexLayout justifyContent="space-between" alignItems="center" mt="md">
               <ForgotPasswordTrigger />
               <Button type="submit" variant={loading ? 'loader' : undefined}>
                 <T id="signIn.button" />
               </Button>
-            </FlexLayout> */}
+            </FlexLayout>
 
-            {/* <FormErrorSummary errors={error && [error.message]} /> */}
+            <FormErrorSummary errors={error && [error.message]} />
           </Form>
         )}
       </Formik>
