+++ packages/falcon-uikit/package.json
@@ -27,6 +27,7 @@
     "@deity/falcon-ui": "^0.2.2",
     "formik": "^1.3.1",
     "graphql-tag": "^2.10.1",
+    "react": "^16.6.3",
     "react-apollo": "2.5.6"
   },
   "devDependencies": {
