+++ packages/falcon-uikit/src/Forms/Field.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Field as FormikField, FieldProps as FormikFieldProps, FieldConfig, getIn } from 'formik';
 import { I18n } from '@deity/falcon-i18n';
 import { Validator } from './validators';
-import { FormContext } from './Form';
+import { FormContext } from './FormContext';
 
 type ValidateSequentially = { (validators: Validator[], label: string): FieldConfig['validate'] };
 const validateSequentially: ValidateSequentially = (validators, label) => value => {
