+++ packages/falcon-uikit/src/Forms/FormContext.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export type FormContextValue = {
+  id: string;
+  name?: string;
+  i18nId?: string;
+};
+
+export const FormContext = React.createContext<FormContextValue>({} as any);
