+++ packages/falcon-uikit/src/Forms/FormFieldLabel.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Label, themed } from '@deity/falcon-ui';
+import { FormFieldArea } from './FormFieldLayout';
+
+export const FormFieldLabel = themed({
+  tag: Label,
+  defaultProps: {
+    gridArea: FormFieldArea.label
+  }
+});
